move requestHook in service folder
