+++ client/src/components/Kanban/KanbanFormComponent.vue
@@ -3,7 +3,7 @@ import { ref, watch, computed, onMounted } from 'vue';
 import Quill from 'quill';
 
 import { client } from '@/services/requestMaker.js';
-import { hookApi } from "@/utils/requestHook.js";
+import { hookApi } from "@/services/requestHook.js";
 import useFormErrors from "@/utils/handleFormErrors.js";
 import logger from "@/utils/logger.js";
 import 'quill/dist/quill.snow.css';
