+++ client/src/components/ToDoList/ToDoItemFormComponent.vue
@@ -3,7 +3,7 @@ import { ref, watch, computed, onMounted } from 'vue';
 import { useRoute } from 'vue-router';
 
 import { client } from '@/services/requestMaker.js';
-import { hookApi } from "@/utils/requestHook.js";
+import { hookApi } from "@/services/requestHook.js";
 import logger from "@/utils/logger.js";
 import useFormErrors from "@/utils/handleFormErrors.js";
