+++ client/src/components/User/UserFormComponent.vue
@@ -2,7 +2,7 @@
 import { ref, watch } from 'vue';
 
 import { client } from '@/services/requestMaker.js';
-import { hookApi } from "@/utils/requestHook.js";
+import { hookApi } from "@/services/requestHook.js";
 import useFormErrors from "@/utils/handleFormErrors.js";
 import logger from "@/utils/logger.js";
