+++ client/src/views/Admin/PriorityView.vue
@@ -2,7 +2,7 @@
 import { ref, onMounted } from 'vue';
 
 import { client } from '@/services/requestMaker.js';
-import { hookApi } from "@/utils/requestHook.js";
+import { hookApi } from "@/services/requestHook.js";
 import logger from "@/utils/logger.js";
 import LoaderComponent from "@/components/LoaderComponent.vue";
 import ModalComponent from "@/components/ModalComponent.vue";
