+++ client/src/views/User/UserDetailView.vue
@@ -2,7 +2,7 @@
 import { ref, onMounted } from 'vue';
 
 import { apiBaseUrl, client } from '@/services/requestMaker.js';
-import { hookApi } from '@/utils/requestHook.js';
+import { hookApi } from '@/services/requestHook.js';
 import { setTitle, setDescription } from "@/utils/documentInfos.js";
 import logger from '@/utils/logger.js';
 import UserFormComponent from "@/components/User/UserFormComponent.vue";
